move countries list into own file
