+++ app/src/test/contrib/components/Responsibilities/Expectations/ExpectationsDescriptionTests.tsx
@@ -4,11 +4,8 @@ import {expect} from "chai";
 
 import {ExpectationMapping} from "../../../../../main/shared/models/Generated";
 import {mockCountry, mockExpectations} from "../../../../mocks/mockModels";
-import {
-    CountriesList,
-    ExpectationsDescription
-} from "../../../../../main/contrib/components/Responsibilities/Expectations/ExpectationsDescription";
-import {Popover, PopoverBody} from "reactstrap";
+import {ExpectationsDescription} from "../../../../../main/contrib/components/Responsibilities/Expectations/ExpectationsDescription";
+import {CountriesList} from "../../../../../main/contrib/components/Responsibilities/Expectations/CountriesList";
 
 describe("ExpectationsDescription", () => {
     it("renders applicable scenarios", () => {
@@ -75,21 +72,4 @@ describe("ExpectationsDescription", () => {
         expect(rendered.find(CountriesList).prop("countries")).to.have.members(countries)
     });
 
-});
-
-describe("Countries list popover", () => {
-    it("renders popover with link", () => {
-        const countries = [mockCountry({name: "countrya"}), mockCountry({name: "countryb"})];
-
-        const rendered = shallow(<CountriesList countries={countries} targetKey={"c1"}/>);
-        expect(rendered.find("#c1").text()).to.equal("view list");
-        expect(rendered.find(Popover).prop("target")).to.equal("c1");
-    });
-
-    it("renders country list in popover", () => {
-        const countries = [mockCountry({name: "countrya"}), mockCountry({name: "countryb"})];
-
-        const rendered = shallow(<CountriesList countries={countries} targetKey={"c1"}/>);
-        expect(rendered.find(PopoverBody).childAt(0).text()).to.equal("countrya, countryb");
-    });
 });
